utils api moved to services file
